Split summary classes
