+++ app/questionnaire/__init__.py
@@ -1,0 +1,1 @@
+from .questionnaire_schema import QuestionnaireSchema
