+++ app/views/contexts/__init__.py
@@ -1,14 +1,17 @@
 from .calculated_summary_context import CalculatedSummaryContext
 from .context import Context
+from .custom_section_summary_context import CustomSectionSummaryContext
+from .final_summary_context import FinalSummaryContext
 from .hub_context import HubContext
 from .list_context import ListContext
-from .summary_context import SummaryContext
-
+from .section_summary_context import SectionSummaryContext
 
 __all__ = [
     "CalculatedSummaryContext",
     "Context",
+    "CustomSectionSummaryContext",
+    "FinalSummaryContext",
     "HubContext",
     "ListContext",
-    "SummaryContext",
+    "SectionSummaryContext",
 ]
