+++ app/views/handlers/section_summary.py
@@ -1,5 +1,5 @@
 from app.views.handlers.content import Content
-from app.views.contexts.summary_context import SummaryContext
+from app.views.contexts import CustomSectionSummaryContext, SectionSummaryContext
 
 
 class SectionSummary(Content):
@@ -8,7 +8,14 @@ class SectionSummary(Content):
         return self._render_block(self.block["id"])
 
     def get_context(self):
-        summary_context = SummaryContext(
+        section_id = self._schema.get_section_id_for_block_id(self.block["id"])
+
+        if not self._schema.get_summary_for_section(section_id):
+            section_summary_context = SectionSummaryContext
+        else:
+            section_summary_context = CustomSectionSummaryContext
+
+        section_summary_context = section_summary_context(
             self._language,
             self._schema,
             self._questionnaire_store.answer_store,
@@ -17,11 +24,4 @@ class SectionSummary(Content):
             self._questionnaire_store.metadata,
         )
 
-        section_id = self._schema.get_section_id_for_block_id(self.block["id"])
-        custom_section_summary = self._schema.get_summary_for_section(section_id)
-
-        return (
-            summary_context.section_summary(self._current_location)
-            if not custom_section_summary
-            else summary_context.custom_section_summary(self._current_location)
-        )
+        return section_summary_context(self._current_location)
